requirements.txt
Django==3.0.8
requests==2.24.0
certbot==1.6.0
django-cors-headers==3.4.0
djangorestframework==3.11.0